chore: code reorg
